requirements.txt
# requirements.txt
appium-python-client>=2.0.0
selenium>=4.0.0